+++ src/Helper/GatewayHelper.php
@@ -29,8 +29,10 @@ use MultiSafepay\Shopware6\PaymentMethods\Fietsenbon;
 use MultiSafepay\Shopware6\PaymentMethods\Gezondheidsbon;
 use MultiSafepay\Shopware6\PaymentMethods\Giropay;
 use MultiSafepay\Shopware6\PaymentMethods\GivaCard;
+use MultiSafepay\Shopware6\PaymentMethods\Good4fun;
 use MultiSafepay\Shopware6\PaymentMethods\Goodcard;
 use MultiSafepay\Shopware6\PaymentMethods\Ideal;
+use MultiSafepay\Shopware6\PaymentMethods\In3;
 use MultiSafepay\Shopware6\PaymentMethods\IngHomePay;
 use MultiSafepay\Shopware6\PaymentMethods\Kbc;
 use MultiSafepay\Shopware6\PaymentMethods\Klarna;
@@ -97,8 +99,10 @@ class GatewayHelper
         Gezondheidsbon::class,
         Giropay::class,
         GivaCard::class,
+        Good4fun::class,
         Goodcard::class,
         Ideal::class,
+        In3::class,
         IngHomePay::class,
         Kbc::class,
         Klarna::class,