+++ src/PaymentMethods/Good4fun.php
@@ -5,9 +5,9 @@
  */
 namespace MultiSafepay\Shopware6\PaymentMethods;
 
-use MultiSafepay\Shopware6\Handlers\KlarnaPaymentHandler;
+use MultiSafepay\Shopware6\Handlers\Good4funPaymentHandler;
 
-class Klarna implements PaymentMethodInterface
+class Good4fun implements PaymentMethodInterface
 {
     /**
      * {@inheritDoc}
@@ -16,7 +16,7 @@ class Klarna implements PaymentMethodInterface
      */
     public function getName(): string
     {
-        return 'Klarna - Buy now, pay later';
+        return 'Good4fun Giftcard';
     }
 
     /**
@@ -36,17 +36,17 @@ class Klarna implements PaymentMethodInterface
      */
     public function getPaymentHandler(): string
     {
-        return KlarnaPaymentHandler::class;
+        return Good4funPaymentHandler::class;
     }
 
     /**
      * {@inheritDoc}
      *
-     * @return string|null
+     * @return string
      */
     public function getGatewayCode(): string
     {
-        return 'KLARNA';
+        return 'GOOD4FUN';
     }
 
     /**
@@ -66,7 +66,7 @@ class Klarna implements PaymentMethodInterface
      */
     public function getMedia(): string
     {
-        return __DIR__  . '/../Resources/views/storefront/multisafepay/logo/klarna.png';
+        return __DIR__  . '/../Resources/views/storefront/multisafepay/logo/good4fun.png';
     }
 
     /**
@@ -79,7 +79,7 @@ class Klarna implements PaymentMethodInterface
         return [
             'de-DE' => [
                 'name'        => $this->getName(),
-                'description' => 'Bezahlen mit ' . $this->getName(),
+                'description' => 'Bezahlen mit '. $this->getName(),
             ],
             'en-GB' => [
                 'name'        => $this->getName(),