+++ src/Resources/app/administration/src/main.js
@@ -1,6 +1,7 @@
 import { Component, Application } from 'src/core/shopware';
 import './components/multisafepay-refund'
 import './components/multisafepay-verify-api-key'
+import './components/multisafepay-support'
 import template from './extension/sw-order-detail/sw-order-detail.html.twig'
 import MultiSafepayApiService from './service/MultiSafepayApi.service'